fix: 버그 수정
- 메뉴판 삭제시 메뉴판 인덱스 조정

+++ src/main/java/com/ourmenu/backend/domain/menu/application/MenuFolderService.java
@@ -52,10 +52,13 @@ public class MenuFolderService {
         menuFolderRepository.delete(menuFolder);
         int maxIndex = menuFolderRepository.findMaxIndex();
 
-        if (menuFolderId != maxIndex) {
-            menuFolderRepository.decrementIndexes(userId, (int) (menuFolderId + 1), maxIndex);
+        if (menuFolder.getIndex() != maxIndex) {
+            menuFolderRepository.decrementIndexes(userId, (menuFolder.getIndex() + 1), maxIndex);
         }
-        awsS3Service.deleteFileAsync(menuFolder.getImgUrl());
+
+        if (menuFolder.getImgUrl() != null) {
+            awsS3Service.deleteFileAsync(menuFolder.getImgUrl());
+        }
     }
 
     @Transactional
